refactor: renomeia diretorio e move componente SearchMenu
